refactor: rename Go module from fraud-operator to fraud
Rename go.miloapis.com/fraud-operator to go.miloapis.com/fraud to match
the repository name.
