+++ internal/controller/fraudevaluation_controller.go
@@ -34,9 +34,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
-	fraudv1alpha1 "go.miloapis.com/fraud-operator/api/v1alpha1"
-	"go.miloapis.com/fraud-operator/internal/datasource"
-	"go.miloapis.com/fraud-operator/internal/provider"
+	fraudv1alpha1 "go.miloapis.com/fraud/api/v1alpha1"
+	"go.miloapis.com/fraud/internal/datasource"
+	"go.miloapis.com/fraud/internal/provider"
 )
 
 const (
