+++ internal/controller/fraudevaluation_controller_test.go
@@ -27,8 +27,8 @@ import (
 	"k8s.io/client-go/tools/events"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	fraudv1alpha1 "go.miloapis.com/fraud-operator/api/v1alpha1"
-	"go.miloapis.com/fraud-operator/internal/provider"
+	fraudv1alpha1 "go.miloapis.com/fraud/api/v1alpha1"
+	"go.miloapis.com/fraud/internal/provider"
 )
 
 // mockProvider is a test implementation of provider.Provider.
