+++ internal/controller/fraudprovider_controller.go
@@ -31,9 +31,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
-	fraudv1alpha1 "go.miloapis.com/fraud-operator/api/v1alpha1"
-	"go.miloapis.com/fraud-operator/internal/provider"
-	"go.miloapis.com/fraud-operator/internal/provider/maxmind"
+	fraudv1alpha1 "go.miloapis.com/fraud/api/v1alpha1"
+	"go.miloapis.com/fraud/internal/provider"
+	"go.miloapis.com/fraud/internal/provider/maxmind"
 )
 
 // supportedProviderTypes lists the valid provider type values.
