+++ internal/controller/suite_test.go
@@ -32,7 +32,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	fraudv1alpha1 "go.miloapis.com/fraud-operator/api/v1alpha1"
+	fraudv1alpha1 "go.miloapis.com/fraud/api/v1alpha1"
 	// +kubebuilder:scaffold:imports
 )
