+++ test/e2e/e2e_test.go
@@ -30,7 +30,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"go.miloapis.com/fraud-operator/test/utils"
+	"go.miloapis.com/fraud/test/utils"
 )
 
 // namespace where the project is deployed in
